Eliminación de copia cuando se hace unswap

+++ src/main/java/com/vrj/coh/tsp/LoteResponse.java
@@ -1,15 +1,13 @@
 package com.vrj.coh.tsp;
 
-import java.math.BigDecimal;
-
 import lombok.Data;
 
 @Data
 public class LoteResponse {
-    private BigDecimal promedio;
+    private double promedio;
     private Solution solution;
 
-    public LoteResponse(BigDecimal promedio, Solution tsp) {
+    public LoteResponse(double promedio, Solution tsp) {
         this.promedio = promedio;
         this.solution = tsp;
     }
